Posture Analyses
By analysing the posture, it receives the user's posture information live from the camera. The received data is processed in the artificial intelligence model and the movement is tracked. If the movement is done correctly, the counter is increased by one, allowing the user to count the number of repetitions.